+++ HUDIYS/src/main/java/hud/iys/dao/IMevzuatDAO.java
@@ -12,9 +12,9 @@ public interface IMevzuatDAO {
 	
 	 public void deleteMevzuat(Mevzuat mevzuat);
 	
-	 public Mevzuat getMevzuatById(int id);
+	 public Mevzuat getMevzuatById(Long id);
 	
 	 public List<Mevzuat> getMevzuatlar();
 	 
-	 public List<Mevzuat> getMevzuatlarByMevzuatSetiId(int mevzuatSetiId);
+	 public List<Mevzuat> getMevzuatlarByMevzuatSetiId(Long mevzuatSetiId);
 }
