+++ HUDIYS/src/main/java/hud/iys/dao/IMevzuatSetiDAO.java
@@ -12,7 +12,7 @@ public interface IMevzuatSetiDAO {
 	
 	 public void deleteMevzuatSeti(MevzuatSeti mevzuatSeti);
 	
-	 public MevzuatSeti getMevzuatSetiById(int id);
+	 public MevzuatSeti getMevzuatSetiById(Long id);
 	
 	 public List<MevzuatSeti> getMevzuatSetleri();
 }
