+++ HUDIYS/src/main/java/hud/iys/service/IKanunIcerikService.java
@@ -14,9 +14,9 @@ public interface IKanunIcerikService {
 	
 	 public void deleteKanunIcerik(KanunIcerik kanunIcerik);
 	
-	 public KanunIcerik getKanunIcerikById(int id);
+	 public KanunIcerik getKanunIcerikById(Long id);
 	
 	 public List<KanunIcerik> getKanunIcerikleri();
 	 
-	 public List<KanunIcerik> getKanunIcerikleriByKanunId(int kanunId);
+	 public List<KanunIcerik> getKanunIcerikleriByKanunId(Long kanunId);
 }
