+++ HUDIYS/src/main/java/hud/iys/service/IParagrafService.java
@@ -13,9 +13,9 @@ public interface IParagrafService {
 	
 	 public void deleteParagraf(Paragraf paragraf);
 	
-	 public Paragraf getParagrafById(int id);
+	 public Paragraf getParagrafById(Long id);
 	
 	 public List<Paragraf> getParagraflar();
 	 
-	 public List<Paragraf> getParagraflarByKanunIcerikId(int kanunIcerikId);
+	 public List<Paragraf> getParagraflarByKanunIcerikId(Long kanunIcerikId);
 }
